requirements.txt
streamlit==0.78.0
Pillow==8.1.0
nltk==3.5